[Ionut] - Forgot to commit color range normalization related to the last commit's changes - Changed fontstash to use texture storage instead of texture image. Removed all traces of textureImage (thank God!) - More WarScene AI work
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@418 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 479cfb0cc4ae8d768b2ab7baff265f384cf89760 [formerly a59dcc209b792fc45664b7d36287e75ab045b7c1] [formerly 7eec27ebc7b2e1c98515a4da2bed471978d0b657 [formerly d604e65cd11eafd8484b638387b84ce4e4767a73]]
Former-commit-id: d97f1e50a715072187b0bb6875c6edd0a39e82cb [formerly 0f012d3cad7d8faf71b99ed08c1bad20aa2af597]
Former-commit-id: d2c47b05f7b0bd60ad3522382a61f61896528785

+++ assets/shaders/GLSL/particles.glsl
@@ -3,7 +3,7 @@
 layout(location = 13) in vec4 particleNormalData;
 // Output data ; will be interpolated for each fragment.
 out vec2 texCoord;
-out vec4 particleColor;
+out uvec4 particleColor;
 out vec4 vertexVP;
 
 // Values that stay constant for the whole mesh.
@@ -44,7 +44,7 @@ void main(){
 
 // Interpolated values from the vertex shaders
 in vec2 texCoord;
-in vec4 particleColor;
+in uvec4 particleColor;
 in vec4 vertexVP;
 
 // Ouput data
